REFACTOR: Created container for daily picture
